feat: Move jest config into package.json to be consitent
