--- packages/minimal/jest.config.js
@@ -1,12 +1,0 @@
-module.exports = {
-  collectCoverage: true,
-  transform: { '^.+\\.ts$': 'ts-jest' },
-  moduleFileExtensions: ['js', 'ts'],
-  testEnvironment: 'node',
-  testMatch: ['**/*.test.ts'],
-  globals: {
-    'ts-jest': {
-      tsConfigFile: './tsconfig.json',
-    },
-  },
-};
