+++ packages/utils/.npmignore
@@ -2,3 +2,4 @@
 !*.js.map
 !*.d.ts
 !*.js
+jest.config.js
