+++ packages/utils/test/async.test.ts
@@ -1,4 +1,4 @@
-import { filterAsync, forget } from '../async';
+import { filterAsync, forget } from '../src/async';
 
 describe('forget', () => {
   const console = {
