+++ packages/utils/test/store.test.ts
@@ -3,7 +3,7 @@ import * as fs from 'fs';
 import * as os from 'os';
 import * as path from 'path';
 
-import { Store } from '../store';
+import { Store } from '../src/store';
 
 jest.useFakeTimers();
